--- src/Example Data/Example Data.ttl
@@ -1,58 +1,0 @@
-@prefix minecraft: <http://example.org/minecraft#> .
-@prefix nodeOntology: <http://example.org/NodeOntology#> .
-
-# Instances of Items
-minecraft:Stick_Plank_made_Instance a minecraft:Stick ;
-	nodeOntology:hasID 1 ;
-    minecraft:obtainedBy minecraft:Stick_Planks_recipe_Instance .
-
-minecraft:Stick_Bamboo_made_Instance a minecraft:Stick ;
-	nodeOntology:hasID 2 ;
-    minecraft:obtainedBy minecraft:Stick_bamboo_recipe_Instance .
-
-minecraft:Cobblestone_Bob a minecraft:Cobblestone ;
-	nodeOntology:hasID 3 .
-
-minecraft:Pickaxe_Instance_Henry a minecraft:Pickaxe ;
-	nodeOntology:hasID 4 ;
-    minecraft:obtainedBy minecraft:PickaxeRecipe_Instance .
-	
-minecraft:Pickaxe_Instance_Gustav a minecraft:Pickaxe ;
-	nodeOntology:hasID 5 .
-
-nodeOntology:Server_b a nodeOntology:Server ;
-	nodeOntology:hasIP "9.2.3.5" .
-
-minecraft:Bamboo_Instance a nodeOntology:NodeFalse ;
-	nodeOntology:hasID 6 ;
-	nodeOntology:pointsToServer minecraft:Server_b . 
-
-minecraft:Plank_Instance a minecraft:Plank ;
-	nodeOntology:hasID 7 .
-
-# Crafting Station Instance
-minecraft:CraftingTable_Instance a minecraft:CraftingTable ;
-	nodeOntology:hasID 8 .
-
-# Recipe Instance: Pickaxe_From_Stick_And_Stone_Recipe 
-minecraft:PickaxeRecipe_Instance a minecraft:Pickaxe_From_Stick_And_Stone_Recipe ;
-	nodeOntology:hasID 9 ;
-    minecraft:hasInput minecraft:Stick_Plank_made ;
-    minecraft:hasInput minecraft:Cobblestone_Bob ;
-    minecraft:hasOutput minecraft:Pickaxe_Instance ;
-    minecraft:usedInStation minecraft:CraftingTable_Instance .
-
-# Recipe Instance: Stick_recipie_bamboo
-minecraft:Stick_bamboo_recipe_Instance a minecraft:Stick_From_Bamboo_Recipe ;
-	nodeOntology:hasID 10 ;
-    minecraft:hasInput minecraft:Bamboo_Instance ;
-    minecraft:hasOutput minecraft:Stick_Bamboo_made ;
-    minecraft:usedInStation minecraft:CraftingTable_Instance .
-
-# Recipe Instance: Stick_recipe_Planks
-minecraft:Stick_Planks_recipe_Instance a minecraft:Stick_From_Planks_Recipe ;
-	nodeOntology:hasID 11 ;
-    minecraft:hasInput minecraft:Stick_Instance1, minecraft:Stick_Instance2 ;
-    minecraft:hasInput minecraft:Cobblestone_Instance1, minecraft:Cobblestone_Instance2 ;
-    minecraft:hasOutput minecraft:Pickaxe_Instance ;
-    minecraft:usedInStation minecraft:CraftingTable_Instance .
