--- src/Example Data/Test Code.txt
@@ -1,1 +1,0 @@
-PREFIX mydb: <http://mydb.org/>PREFIX mine: <http://example.org/minecraft#>SELECT ?recipe ?stationWHERE{     ?recipe a mine:Recipe ;    	mine:hasOutput mine:Stick}
