--- src/go.mod
@@ -1,1 +1,0 @@
-module github.com/deiu/rdf2go
