--- src/serverInfo.txt
@@ -1,3 +1,0 @@
-"a" "b"
-["sticks", "planks"]
-["b","c"]
